# release-2.1: opt: Reuse Memo data structure