Backport 1/1 commits from #29573.

/cc @cockroachdb/release

---

Embed the Memo data structure in Factory and allow it to be reused.
This gives a sizeable speedup on simple queries:
```
name                      old time/op  new time/op  delta
Phases/kv-read/Explore    26.2µs ± 2%  23.3µs ± 1%  -11.02%  (p=0.008 n=5+5)
Phases/planning1/Explore  10.6µs ± 1%   8.5µs ± 0%  -19.96%  (p=0.008 n=5+5)
Phases/planning2/Explore  31.9µs ± 2%  28.7µs ± 4%  -10.12%  (p=0.008 n=5+5)
Phases/planning3/Explore  34.6µs ± 2%  31.9µs ± 4%   -7.87%  (p=0.008 n=5+5)
Phases/planning4/Explore  42.1µs ± 2%  38.6µs ± 2%   -8.37%  (p=0.008 n=5+5)
Phases/planning5/Explore  34.0µs ± 1%  30.8µs ± 1%   -9.44%  (p=0.008 n=5+5)
Phases/planning6/Explore  57.8µs ± 2%  54.3µs ± 2%   -6.11%  (p=0.008 n=5+5)
```

Release note: None